papiea-sdk/python/tests/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
idna==2.10
more-itertools==8.4.0
multidict==4.7.6
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
pytest-asyncio==0.14.0
PyYAML==5.3.1
six==1.15.0
wcwidth==0.2.5
yarl==1.4.2
